pyproject.toml
[project]
name = "ttcs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "pdfplumber",
    "pymupdf",
    "pypdf",
    "python-docx",
    "docling",
    "langchain-text-splitters>=0.3.9",
    "onnxruntime>=1.14.0",
    "onnxruntime-gpu>=1.24.2",
    "transformers>=4.40.0",
    "sentencepiece",
    "safetensors",
    "numpy",
    "torch>=2.10.0",
    "pydantic>=2.10.0",
    "chromadb>=1.5.2",
]

[tool.setuptools.packages.find]
where = ["."]
